centralize testing mocks and remove custom test clock when k8s test util will do
